Separating cve unit tests by endpoint and positive/negative tests to improve readability and maintenance
